% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{angle_from_2pts}
\alias{angle_from_2pts}
\title{Get angle from 2 points}
\usage{
angle_from_2pts(coords, pt1 = 1, pt2 = 2)
}
\arguments{
\item{coords}{The coordinate array}

\item{pt1}{The first point}

\item{pt2}{The second point}
}
\value{
double of angle in radians
}
\description{
Get angle from 2 points
}
\keyword{internal}
