% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimlist.R
\name{as_stimlist}
\alias{as_stimlist}
\title{Convert list to stimlist}
\usage{
as_stimlist(x, tem = FALSE)
}
\arguments{
\item{x}{The object}

\item{tem}{require templates (images without templates are removed)}
}
\value{
A stimlist
}
\description{
Checks if an object is a list of class stimlist. If class stim, wrap in stimlist. If a properly structured list without the right class, add the right class.
}
