% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg.R
\name{avg}
\alias{avg}
\title{Average Images}
\usage{
avg(
  stimuli,
  texture = TRUE,
  norm = c("none", "twopoint", "rigid"),
  normpoint = 0:1
)
}
\arguments{
\item{stimuli}{list of stimuli to average}

\item{texture}{logical; whether textured should be averaged}

\item{norm}{how to normalise}

\item{normpoint}{points for twopoint normalisation}
}
\value{
list of stimuli with the average image and template
}
\description{
Average Images
}
\examples{
\dontrun{
  avg <- demo_stim() |> avg()
}
}
\seealso{
WebMorph.org functions
\code{\link{continuum}()},
\code{\link{loop}()},
\code{\link{symmetrize}()},
\code{\link{trans}()}
}
\concept{webmorph}
