% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank.R
\name{blank}
\alias{blank}
\title{Make blank images}
\usage{
blank(n = 1, width = 100, height = 100, color = "white", names = "img")
}
\arguments{
\item{n}{the number of images to return}

\item{width, height}{image dimensions}

\item{color}{background color, see \code{\link[=color_conv]{color_conv()}}}

\item{names}{names of the images (appended with index if < n)}
}
\value{
list of stimuli
}
\description{
Make blank images
}
\examples{
stimuli <- blank(5, 100, 250, color = rainbow(5))

label(stimuli, size = 20)
}
\seealso{
Stimulus creating functions
\code{\link{animate}()},
\code{\link{as_stimlist}()},
\code{\link{new_stimlist}()},
\code{\link{new_stim}()},
\code{\link{read_img}()},
\code{\link{read_stim}()},
\code{\link{read_tem}()},
\code{\link{write_stim}()}
}
\concept{stim}
