% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{bounds}
\alias{bounds}
\title{Get template bounds}
\usage{
bounds(stimuli, each = FALSE)
}
\arguments{
\item{stimuli}{A stimlist}

\item{each}{Whether to calculate max and min for the full set (default) or each image separately}
}
\value{
A list of min and max x and y values
}
\description{
Get template bounds
}
\examples{
demo_stim() \%>\% bounds() \%>\% str()

demo_stim() \%>\% bounds(each = TRUE)
}
