% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c.stim}
\alias{c.stim}
\title{Combine stim}
\usage{
\method{c}{stim}(...)
}
\arguments{
\item{...}{stim to be concatenated}
}
\value{
stimlist
}
\description{
Combine stim
}
\keyword{internal}
\keyword{methods}
