% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c.stimlist}
\alias{c.stimlist}
\title{Combine stimlists}
\usage{
\method{c}{stimlist}(...)
}
\arguments{
\item{...}{stimlists to be concatenated}
}
\value{
stimlist
}
\description{
Combine stimlists
}
\keyword{internal}
