% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{centroid}
\alias{centroid}
\title{Get center coordinates}
\usage{
centroid(stimuli, points = NULL)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{points}{which points to include (0-based); if NULL, all points will be used}
}
\value{
named matrix of centroid x and y coordinates
}
\description{
Get center coordinates
}
\examples{
demo_stim() |> centroid()

# get the centre of the eye points
demo_stim() |> centroid(0:1)
}
