% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tem.R
\name{convert_tem}
\alias{convert_tem}
\title{Convert templates}
\usage{
convert_tem(
  stimuli,
  from = c("guess", "frl", "fpp106", "fpp83", "dlib70", "dlib7"),
  to = c("frl", "fpp106", "fpp83", "dlib70", "dlib7")
)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{from}{id of template definition of stimlist images}

\item{to}{id of template definition to convert to}
}
\value{
stimlist with converted templates
}
\description{
Convert templates
}
\examples{
fpp106 <- demo_stim() |> convert_tem("guess", "fpp106")
}
