% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop_pad}
\alias{crop_pad}
\title{Crop to template boundaries and pad}
\usage{
crop_pad(
  stimuli,
  top = 10,
  right = top,
  bottom = top,
  left = right,
  each = FALSE,
  ...
)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{top}{number of pixels to pad the top}

\item{right}{number of pixels to pad the right side}

\item{bottom}{number of pixels to pad the bottom}

\item{left}{number of pixels to pad the left side}

\item{each}{Whether to calculate bounds for the full set (default) or each image separately}

\item{...}{additional arguments to pass to \code{\link{crop}}}
}
\value{
list of class stimlist
}
\description{
Calculate the max and min x and y coordinates across the stimuli and crop all image to this plus padding
}
\examples{
cpad <- demo_stim() \%>\% crop_pad(20) \%>\% draw_tem()
plot(cpad)
}
