% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delin.R
\name{delin}
\alias{delin}
\title{Manually delineate images}
\usage{
delin(stimuli)
}
\arguments{
\item{stimuli}{list of class stimlist}
}
\value{
stimlist with new templates
}
\description{
Adjust the templates in a shiny interface. This will overwrite existing templates.
}
\examples{
\dontrun{
# adjust existing delineations
stimuli <- demo_stim() |> delin()

# create new delineations from scratch
stimuli <- demo_stim() |> remove_tems() |> delin()
}
}
