% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_stim.R
\name{demo_stim}
\alias{demo_stim}
\alias{demo_tems}
\title{Demo Stimuli}
\usage{
demo_stim(pattern = NULL)

demo_tems(pattern = NULL)
}
\arguments{
\item{pattern}{defaults to all files}
}
\value{
stimlist
}
\description{
A convenience function to get demo stimuli
}
\examples{
demo_stim() |> plot()

}
