% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_tem.R
\name{draw_tem}
\alias{draw_tem}
\title{Draw template}
\usage{
draw_tem(
  stimuli,
  pt.color = wm_opts("pt.color"),
  pt.alpha = 0.75,
  pt.size = NULL,
  pt.shape = c("circle", "cross", "index"),
  line.color = wm_opts("line.color"),
  line.alpha = 0.5,
  line.size = NULL,
  bg = "image"
)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{pt.color, line.color}{line or point color}

\item{pt.alpha, line.alpha}{transparency (0-1), ignored if color is a hex value with transparency}

\item{pt.size, line.size}{size in pixels (scales to image if NULL)}

\item{pt.shape}{the shape of the points ("circle", "cross", "index")}

\item{bg}{background color ("image" uses original image)}
}
\value{
stimlist with template images
}
\description{
Set alpha to 0 to omit lines or points.
}
\examples{
demo_stim() \%>\% draw_tem() \%>\% plot()
}
