% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{features}
\alias{features}
\title{Feature Points}
\usage{
features(..., tem_id = "FRL")
}
\arguments{
\item{...}{a vector of feature names (see Details)}

\item{tem_id}{template ID (currently only works for FRL)}
}
\value{
vector of corresponding FRL template indices
}
\description{
Get point indices for features, usually for use with \code{\link{subset_tem}}.
}
\details{
Available features for the FRL template are: "gmm", "oval", "face", "mouth", "nose", "eyes", "brows", "left_eye",  "right_eye", "left_brow",  "right_brow", "ears", "undereyes", "teeth", "smile_lines", "cheekbones", "philtrum", "chin", "neck", "halo".
}
\examples{
features("mouth")
features("gmm")

}
