% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_size}
\alias{format_size}
\title{Format file size}
\usage{
format_size(x)
}
\arguments{
\item{x}{the file size in bytes}
}
\value{
human-readable file size
}
\description{
Format file size
}
\examples{
format_size(1024*1024)
}
\keyword{internal}
