% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_delete.R
\name{frl_features}
\alias{frl_features}
\title{Get point indices for FRL features}
\usage{
frl_features(...)
}
\arguments{
\item{...}{a vector of feature names or "imprecise" to choose undereyes, ears, halo, mouth2, smile_lines, cheekbones, philtrum, chin, neck points}
}
\value{
vector of corresponding FRL template indices
}
\description{
Get point indices for FRL features
}
\examples{
frl_features("mouth")
frl_features("left_eye", "right_eye")

}
