% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_imgs}
\alias{get_imgs}
\title{Get Images into List}
\usage{
get_imgs(stimuli)
}
\arguments{
\item{stimuli}{list of stimuli}
}
\value{
list of magick images
}
\description{
Get Images into List
}
\keyword{internal}
