% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_func.R
\name{greyscale}
\alias{greyscale}
\alias{grayscale}
\title{Make images greyscale}
\usage{
greyscale(stimuli)

grayscale(stimuli)
}
\arguments{
\item{stimuli}{list of class stimuli}
}
\value{
stimlist with new images
}
\description{
Make images greyscale
}
\examples{
stimuli <- demo_stim()
grey_stim <- greyscale(stimuli)
plot(grey_stim)
stimuli <- demo_stim()
gray_stim <- grayscale(stimuli)
plot(gray_stim)
}
\seealso{
Stimulus manipulation functions
\code{\link{align}()},
\code{\link{crop_tem}()},
\code{\link{crop}()},
\code{\link{horiz_eyes}()},
\code{\link{image_func}()},
\code{\link{mask_oval}()},
\code{\link{mask}()},
\code{\link{mirror}()},
\code{\link{pad}()},
\code{\link{resize}()},
\code{\link{rotate}()},
\code{\link{to_size}()}

Stimulus manipulation functions
\code{\link{align}()},
\code{\link{crop_tem}()},
\code{\link{crop}()},
\code{\link{horiz_eyes}()},
\code{\link{image_func}()},
\code{\link{mask_oval}()},
\code{\link{mask}()},
\code{\link{mirror}()},
\code{\link{pad}()},
\code{\link{resize}()},
\code{\link{rotate}()},
\code{\link{to_size}()}
}
\concept{manipulators}
