% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{height}
\alias{height}
\title{Image heights}
\usage{
height(stimuli, type = c("all", "min", "max", "unique"))
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{type}{whether to return all heights, min, max, or only unique heights}
}
\value{
vector of heights
}
\description{
Image heights
}
\examples{

demo_stim() |> height()
}
