% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{horiz_eyes}
\alias{horiz_eyes}
\title{Make eyes horizontal}
\usage{
horiz_eyes(
  stimuli,
  left_eye = 0,
  right_eye = 1,
  fill = wm_opts("fill"),
  patch = FALSE
)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{left_eye}{The first point to align (defaults to 0)}

\item{right_eye}{The second point to align (defaults to 1)}

\item{fill}{background color to pass to rotate}

\item{patch}{whether to use the patch function to set the background color}
}
\value{
stimlist with rotated tems and/or images
}
\description{
Make eyes horizontal
}
\examples{
demo_stim("lisa") \%>\% horiz_eyes() \%>\% plot()

}
