% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_comp.R
\name{image_comp}
\alias{image_comp}
\title{Image Comparison}
\usage{
image_comp(stim, ref_stim, metric = "MSE", fuzz = 0, scale = FALSE)
}
\arguments{
\item{stim, ref_stim}{Stimuli to compare}

\item{metric}{string with a metric from magick::metric_types(): "Undefined", "AE", "Fuzz", "MAE", "MEPP", "MSE", "NCC", "PAE", "PHASH", "PSNR", "RMSE"}

\item{fuzz}{relative color distance (value between 0 and 100) to be considered similar in the filling algorithm (only useful for AE)}

\item{scale}{whether to scale the values so that the maximum value is 1 and the minimum is 0 (only useful when stim is more than 1 image and includes ref_stim)}
}
\value{
Difference metric
}
\description{
This is just a convenient way to use magick::image_compare with webmorph stimuli. It defaults to the "MSE" metric, which gives a linearly increasing score to images along a morph continuum.
}
\details{
Metric Types
\itemize{
\item Undefined: ?
\item AE: Absolute Error
\item Fuzz: ?
\item MAE: Mean Absolute Error
\item MEPP: Mean Error Per Pixel
\item MSE: Mean Squared Error
\item NCC: Normalized Cross Correlation
\item PAE: Peak Absolute Error
\item PHASH: Perceptual Hash
\item PSNR: Peak Signal-to-Noise Ratio
\item RMSE: Root Mean Squared Error
}

How these metrics behave when comparing a morph continuum to its first image.

Increases with morph distance:
\itemize{
\item very strong negative exponential decay at 0 fuzz; more linear with higher fuzz: AE
\item strong negative exponential decay: PAE
\item slight negative exponential decay: Fuzz, RMSE
\item linear: MAE, MEPP, MSE
\item no idea: PHASH
}

Decreases with morph distance:
\itemize{
\item linear: NCC, Undefined
\item slight exponential decay: PSNR
}
}
\examples{
stimuli <- demo_stim()
image_comp(stimuli, stimuli$m_multi)
image_comp(stimuli, stimuli$m_multi, "AE")
image_comp(stimuli, stimuli$m_multi, "AE", fuzz = 5)
}
