% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Add a label}
\usage{
label(
  stimuli,
  text = TRUE,
  gravity = "north",
  location = "+0+10",
  degrees = 0,
  size = NULL,
  font = "sans",
  style = "normal",
  weight = 400,
  kerning = 0,
  decoration = NULL,
  color = "black",
  strokecolor = NULL,
  boxcolor = NULL
)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{text}{a vector of the label text(s) or TRUE to use stimlist names}

\item{gravity}{string with gravity value from \code{magick::gravity_types}.}

\item{location}{geometry string with location relative to gravity}

\item{degrees}{rotates text around center point}

\item{size}{font size in pixels (if NULL, scales to 5\% image height)}

\item{font}{string with font family such as "sans", "mono", "serif", "Times", "Helvetica", "Trebuchet", "Georgia", "Palatino" or "Comic Sans".}

\item{style}{value of style_types for example "italic"}

\item{weight}{thickness of the font, 400 is normal and 700 is bold.}

\item{kerning}{increases or decreases whitespace between letters}

\item{decoration}{value of decoration_types for example "underline"}

\item{color}{a vector of the label colour(s)}

\item{strokecolor}{adds a stroke (border around the text)}

\item{boxcolor}{adds a background color}
}
\value{
stimlist with labelled images
}
\description{
This is just a wrapper function for magick::image_annotate that allows more flexibility in color input.
}
\examples{

demo_stim("test") \%>\%
  label(c("CHINWE", "GEORGE"), color = "red") \%>\%
  plot()
}
