% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_tbl.R
\name{list_to_tbl}
\alias{list_to_tbl}
\title{List format to table format}
\usage{
list_to_tbl(list, rownames = NULL)
}
\arguments{
\item{list}{a list of lists}

\item{rownames}{whether to return a table with no rownames (NULL), rownames from the list item names (NA), or as a new column (the column name as a string)}
}
\value{
a data table
}
\description{
List format to table format
}
\keyword{internal}
