% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{london_info}
\alias{london_info}
\title{London Set Information}
\format{
A data frame with 102 rows and 4 variables:
\describe{
\item{face_id}{face's ID}
\item{face_gender}{face's gender (female, male)}
\item{face_age}{face's age (19-47 years)}
\item{face_eth}{face's ethnic group (black, east_asian, west_asian, white)}
}
}
\source{
\url{https://figshare.com/articles/dataset/Face_Research_Lab_London_Set/5047666?file=27397184}
}
\usage{
london_info
}
\description{
A dataset containing information on the 102 faces in the Face Research Lab London Set
}
\keyword{datasets}
