% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop.R
\name{loop}
\alias{loop}
\title{Loop}
\usage{
loop(stimuli, steps = 10, ...)
}
\arguments{
\item{stimuli}{list of images to morph between}

\item{steps}{number of steps from one image to the next}

\item{...}{arguments to pass to \code{\link{trans}}}
}
\value{
stimlist
}
\description{
Loop
}
\examples{
\dontrun{
  stimuli <- demo_stim() |> 
    resize(300) |>
    rep(4) |>
    rotate(seq(0, 360-45, 45))
  loop <- loop(stimuli, 5)
  animate(loop, 10)
}
}
