% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{mask}
\alias{mask}
\title{Mask Images with templates}
\usage{
mask(
  stimuli,
  mask = "face",
  fill = wm_opts("fill"),
  reverse = FALSE,
  expand = 1,
  tem_id = "FRL"
)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{mask}{vector of masks or a custom list of template points}

\item{fill}{color to make the mask}

\item{reverse}{if TRUE, the mask covers the listed areas}

\item{expand}{how many pixels to expand the mask}

\item{tem_id}{template ID to pass on to \code{tem_def} to get built-in mask definitions, usually one of "FRL", "fpp106" or "fpp83"}
}
\value{
stimlist with masked images
}
\description{
Create a mask using template points.
}
\details{
For FRL templates, the argument \code{mask} can be a vector with one or more of the following: oval, face, neck, ears (left_ear, right_ear), eyes (left_eye, right_eye), brows (left_brow, right_brow), mouth, teeth, nose.

For Face++ templates (fpp83 or fpp106), the argument \code{mask} can be a vector with one or more of the following: face, eyes (left_eye, right_eye), brows (left_brow, right_brow), mouth, teeth, nose. Because these templates have no forehead points, "face" is usually disappointing.

Set custom masks using the template points (0-based). View an image with labelled templates using \code{plot(stim, pt.plot = TRUE, pt.shape="index")}. Separate points along a line with commas, line segments with semicolons, and mask areas with colons. For example, this would be the custom mask for the eyes in the fpp83 template:

\code{"44,4,56,51,79;79,58,11,25,44:61,67,38,34,40;40,41,17,47,61"}

If you set expand = 0, there is sometimes a thin visible line where multiple components of the mask touch.
}
\examples{
stimuli <- demo_stim()
masked <- mask(stimuli, c("face", "neck", "ears"), "red")

revmasked <- mask(stimuli, "eyes", "#FFFF00", TRUE) \%>\%
  mask("brows", "purple", TRUE) \%>\%
  mask("nose", "#FF000066", TRUE) \%>\%
  mask("mouth", "blue", TRUE)

masked_fpp <- demo_stim("lisa")  \%>\%
  crop_pad() \%>\%
  mask(mask = c("face"), tem_id = "fpp106")

}
