% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimlist.R
\name{new_stimlist}
\alias{new_stimlist}
\title{Make a new stimlist}
\usage{
new_stimlist(..., .names = NULL)
}
\arguments{
\item{...}{Lists with class "stim"}

\item{.names}{Names for each stimulus}
}
\value{
A stimlist
}
\description{
Make a new stimlist
}
\seealso{
Stimulus creating functions
\code{\link{animate}()},
\code{\link{as_stimlist}()},
\code{\link{blank}()},
\code{\link{new_stim}()},
\code{\link{read_img}()},
\code{\link{read_stim}()},
\code{\link{read_tem}()},
\code{\link{write_stim}()}
}
\concept{stim}
\keyword{internal}
