% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.stim}
\alias{plot.stim}
\title{Plot stim}
\usage{
\method{plot}{stim}(x, y, ...)
}
\arguments{
\item{x}{stim}

\item{y}{omitted}

\item{...}{Arguments to be passed to \code{\link{plot_stim}}}
}
\value{
stimlist
}
\description{
Plot stim
}
\examples{
demo_stim("test")[[1]] |> plot()
}
