% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.stimlist}
\alias{plot.stimlist}
\title{Plot stimuli}
\usage{
\method{plot}{stimlist}(x, y, ...)
}
\arguments{
\item{x}{list of class stimlist}

\item{y}{omitted}

\item{...}{Arguments to be passed to \code{\link{plot_stim}}}
}
\value{
stimlist
}
\description{
Plot stimuli in a stimlist.
}
\examples{
demo_stim("test") \%>\% plot()

}
