% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.stim}
\alias{print.stim}
\title{Print stim}
\usage{
\method{print}{stim}(x, ...)
}
\arguments{
\item{x}{a list of class stim}

\item{...}{arguments passed to or from other methods}
}
\value{
prints summary info and returns x
}
\description{
Print stim
}
