% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.stimlist}
\alias{print.stimlist}
\title{Print stimlist}
\usage{
\method{print}{stimlist}(x, ...)
}
\arguments{
\item{x}{list of stimuli}

\item{...}{arguments passed to or from other methods}
}
\value{
prints summary info and returns x
}
\description{
Print stimlist
}
\keyword{internal}
