% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{procrustes_coords}
\alias{procrustes_coords}
\title{Procrustes align templates}
\usage{
procrustes_coords(data, ref_img = NULL)
}
\arguments{
\item{data}{Template points}

\item{ref_img}{Reference image}
}
\value{
coordinates
}
\description{
Procrustes align templates
}
\keyword{internal}
