% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_delete.R
\name{pt_delete}
\alias{pt_delete}
\title{Delete template points}
\usage{
pt_delete(stimuli, ...)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{...}{vectors of points to delete}
}
\value{
stimlist with altered templates
}
\description{
Delete template points
}
\examples{
demo_stim()[1] \%>\%
  pt_delete(frl_features("imprecise")) \%>\%
  draw_tem()

}
