% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_delin.R
\name{quick_delin}
\alias{quick_delin}
\title{Quickly delineate images}
\usage{
quick_delin(stimuli, n_points = 0)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{n_points}{Number of points after which the template automatically saves and moves on to the next image.}
}
\value{
stimlist with new templates
}
\description{
Best used for just a few points, like delineating the eyes, or top, right, bottom and left boundaries. This will overwrite existing templates.
}
