% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stim.R
\name{read_img}
\alias{read_img}
\title{Read image file}
\usage{
read_img(path)
}
\arguments{
\item{imgfile}{}
}
\value{
list of image info
}
\description{
Read image file
}
\examples{
path <- system.file("extdata/test/f_multi.jpg", 
                    package = "webmorphR")
img <- read_img(path)
}
\seealso{
Stimulus creating functions
\code{\link{animate}()},
\code{\link{as_stimlist}()},
\code{\link{blank}()},
\code{\link{new_stimlist}()},
\code{\link{new_stim}()},
\code{\link{read_stim}()},
\code{\link{read_tem}()},
\code{\link{write_stim}()}
}
\concept{stim}
\keyword{internal}
