% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stim.R
\name{read_tem}
\alias{read_tem}
\title{Read tem file}
\usage{
read_tem(path)
}
\arguments{
\item{path}{path to template file}
}
\value{
list of template info
}
\description{
Read tem file
}
\examples{
path <- system.file("extdata/test/f_multi.tem", 
                    package = "webmorphR")
tem <- read_tem(path)
}
\seealso{
Stimulus creating functions
\code{\link{animate}()},
\code{\link{as_stimlist}()},
\code{\link{blank}()},
\code{\link{new_stimlist}()},
\code{\link{new_stim}()},
\code{\link{read_img}()},
\code{\link{read_stim}()},
\code{\link{write_stim}()}
}
\concept{stim}
\keyword{internal}
