% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{remove_tem}
\alias{remove_tem}
\title{Remove templates}
\usage{
remove_tem(stimuli)
}
\arguments{
\item{stimuli}{list of stimuli}
}
\value{
list of stimuli
}
\description{
Remove templates
}
\examples{
demo_stim() |> remove_tem()
}
\seealso{
Template functions
\code{\link{auto_delin}()},
\code{\link{average_tem}()},
\code{\link{centroid}()},
\code{\link{change_lines}()},
\code{\link{delin}()},
\code{\link{draw_tem}()},
\code{\link{features}()},
\code{\link{get_point}()},
\code{\link{require_tems}()},
\code{\link{same_tems}()},
\code{\link{squash_tem}()},
\code{\link{subset_tem}()},
\code{\link{tem_def}()},
\code{\link{viz_tem_def}()}
}
\concept{tem}
