% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_stim.R
\name{rename_stim}
\alias{rename_stim}
\title{Set stimulus names in a stimlist}
\usage{
rename_stim(
  stimuli,
  new_names = NULL,
  prefix = "",
  suffix = "",
  pattern = NULL,
  replacement = NULL,
  ...
)
}
\arguments{
\item{stimuli}{A list of stimuli}

\item{new_names}{Vector of new names - must be the same length as the stimlist}

\item{prefix}{String to prefix to each name}

\item{suffix}{String to append to each name}

\item{pattern}{Pattern for gsub}

\item{replacement}{Replacement for gsub}

\item{...}{Additional arguments to pass on to \code{base::gsub()}}
}
\value{
A list of stimuli with the new names
}
\description{
Set stimulus names in a stimlist
}
\examples{
demo_stim() |>
  rename_stim(prefix = "new_") |>
  names()
}
\seealso{
Other info: 
\code{\link{add_info}()},
\code{\link{compare}()},
\code{\link{get_info}()},
\code{\link{get_point}()},
\code{\link{height}()},
\code{\link{metrics}()},
\code{\link{width}()}
}
\concept{info}
