% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep.stim}
\alias{rep.stim}
\title{Repeat stim in a list}
\usage{
\method{rep}{stim}(x, ...)
}
\arguments{
\item{x}{A list of class stim}

\item{...}{Additional arguments to pass on to \code{base::rep()}}
}
\value{
A stimlist
}
\description{
Repeat stim in a list
}
\examples{
a <- demo_stim()
rep(a[[1]], 3)
}
\keyword{internal}
\keyword{methods}
