% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep.stimlist}
\alias{rep.stimlist}
\title{Repeat stim in a list}
\usage{
\method{rep}{stimlist}(x, ...)
}
\arguments{
\item{x}{A stimlist}

\item{...}{Additional arguments to pass on to \code{base::rep()}}
}
\value{
A stimlist
}
\description{
Repeat stim in a list
}
\examples{
\dontrun{
demo_stim() |>
  rep(3) |>
  rotate(seq(10, 60, 10), fill = rainbow(6)) |>
  plot()
}
}
\keyword{internal}
