% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize.R
\name{resize}
\alias{resize}
\title{Resize templates and images}
\usage{
resize(stimuli, width = NULL, height = NULL)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{width}{new width (in pixels or percent if < 10)}

\item{height}{new height (in pixels or percent if < 10)}
}
\value{
stimlist with resized tems and/or images
}
\description{
Resize templates and images
}
\examples{
resized <- demo_stim() \%>\%
  resize(.5, .75) \%>\%
  draw_tem()

}
