% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotated_size}
\alias{rotated_size}
\title{Image size after rotation}
\usage{
rotated_size(width, height, degrees)
}
\arguments{
\item{width}{Width of the original image}

\item{height}{Height of the original image}

\item{degrees}{Rotation in degreed}
}
\value{
list of rotated width and height
}
\description{
Image size after rotation
}
\keyword{internal}
