% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{same_tems}
\alias{same_tems}
\title{Check All Templates are the Same}
\usage{
same_tems(stimuli)
}
\arguments{
\item{stimuli}{list of class stimlist}
}
\value{
logical
}
\description{
Check All Templates are the Same
}
\examples{
stim <- demo_stim()
stim2 <- subset_tem(stim, features("gmm"))

same_tems(stim)

c(stim, stim2) \%>\% same_tems()
}
