% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setnames.R
\name{setnames}
\alias{setnames}
\title{Set stimulus names in a stimlist}
\usage{
setnames(
  stimuli,
  new_names = NULL,
  prefix = "",
  suffix = "",
  pattern = NULL,
  replacement = NULL,
  ...
)
}
\arguments{
\item{stimuli}{A stimlist}

\item{new_names}{Vector of new names - must be the same length as the stimlist}

\item{prefix}{String to prefix to each name}

\item{suffix}{String to append to each name}

\item{pattern}{Pattern for gsub}

\item{replacement}{Replacement for gsub}

\item{...}{Additional arguments to pass on to \code{base::gsub()}}
}
\value{
a stimlist
}
\description{
Set stimulus names in a stimlist
}
\examples{
demo_stim() |>
  setnames(prefix = "new_") |>
  names()
}
