% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_size.R
\name{social_media_size}
\alias{social_media_size}
\title{Social Media Image Sizes}
\usage{
social_media_size(platform = c("twitter", "instagram"), type = "default")
}
\arguments{
\item{platform}{currently only "twitter"}

\item{type}{which type of image}
}
\value{
named vector of width and height in pixels
}
\description{
A convenience function for getting recommended dimensions for images on social media sites.
}
\details{
Twitter:

link: Image from a Tweet with shared link
one: Tweet sharing a single image (default)
two: Tweet sharing two images
three_left: Tweet sharing three images, Left image
three_tight Tweet sharing three images, Right images
four: Tweet sharing four images

Instagram:

feed_large: (default)
feed_small:
stories_large:
stories_small:
}
\examples{
social_media_size("twitter", "link")
social_media_size("twitter", "one")
social_media_size("twitter", "two")

}
