% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{squash_tem}
\alias{squash_tem}
\title{Squash Template Points}
\usage{
squash_tem(stimuli)
}
\arguments{
\item{stimuli}{list of class stimlist}
}
\value{
list of class stimlist
}
\description{
Move template points that are outside the image boundaries (e.g., negative values or larger than image width or height) to the borders of the image.
}
\examples{
nosquash <- demo_stim()[1] \%>\% 
  crop(0.4, 0.5) \%>\% pad(50) \%>\% 
  draw_tem(line.size = 3)

squashed <- demo_stim()[1] \%>\% 
  crop(0.4, 0.5) \%>\% squash_tem() \%>\% pad(50) \%>\% 
  draw_tem(line.size = 3)
  
# c(nosquash, squashed) \%>\% plot()
}
