% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[.stimlist}
\alias{[.stimlist}
\title{Extract stimlist elements}
\usage{
\method{[}{stimlist}(x, i)
}
\arguments{
\item{x}{stimlist from which to extract elements}

\item{i}{indices to be selected}
}
\value{
stimlist
}
\description{
Extract stimlist elements
}
