% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[[<-.stimlist}
\alias{[[<-.stimlist}
\title{Replace stimlist element}
\usage{
\method{[[}{stimlist}(x, i) <- value
}
\arguments{
\item{x}{stimlist from which to extract elements}

\item{i}{index to be replaced}

\item{value}{stim element to replace with}
}
\value{
stimlist
}
\description{
Replace stimlist element
}
\keyword{internal}
\keyword{methods}
