% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset.stimlist}
\alias{subset.stimlist}
\title{Subset Stimulus Lists}
\usage{
\method{subset}{stimlist}(x, subset, ...)
}
\arguments{
\item{x}{a list of class stimlist}

\item{subset}{a character string to use as a pattern for searching stimulus IDs, or a logical expression indicating elements or rows to keep: missing values are taken as false.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
a list of class stimlist
}
\description{
Returns a subset of the stimulus list meeting the condition.
}
\examples{
f <- demo_stim() \%>\% subset("f_")
}
