% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{subset_tem}
\alias{subset_tem}
\title{Subset template points}
\usage{
subset_tem(stimuli, ..., keep = TRUE)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{...}{vectors of points to keep or delete}

\item{keep}{whether to keep or delete the points}
}
\value{
stimlist with altered templates
}
\description{
Subset template points
}
\examples{
demo_stim()[1] \%>\%
  subset_tem(features("gmm")) \%>\%
  draw_tem()

}
