% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{svgControlPoints}
\alias{svgControlPoints}
\title{Get Control Points}
\usage{
svgControlPoints(x0, y0, x1, y1, x2, y2, t = 0.3)
}
\arguments{
\item{x0, y0}{coordinates of previous point}

\item{x1, y1}{coordinates of main point}

\item{x2, y2}{coordinates of next point}

\item{t}{curving factor (0-1)}
}
\value{
x- and y-coordinates of control points (x1, y1, x2, y2)
}
\description{
Get Control Points
}
\examples{
\dontrun{
svgControlPoints(10, 10, 20, 20, 30, 10)
}
}
\keyword{internal}
