% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{svgCubicTo}
\alias{svgCubicTo}
\title{SVG Path Quadratic Curve}
\usage{
svgCubicTo(x1, y1, x2, y2, x, y, digits = 2)
}
\arguments{
\item{x1, y1, x2, y2}{coordinates of control points}

\item{x, y}{coordinates of main point}

\item{digits}{number of digits to round to}
}
\value{
string with Q path component
}
\description{
SVG Path Quadratic Curve
}
\examples{
\dontrun{
svgCubicTo(20.123, 30.456, 40.123, 50.456, 60.123, 70.456)
}
}
\keyword{internal}
