% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{svgLines}
\alias{svgLines}
\title{Visualise template lines}
\usage{
svgLines(stim)
}
\arguments{
\item{stim}{the stimulus object}
}
\value{
SVG text
}
\description{
Visualise template lines
}
\keyword{internal}
