% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{svgLineTo}
\alias{svgLineTo}
\title{SVG Path Line}
\usage{
svgLineTo(x, y, digits = 2)
}
\arguments{
\item{x, y}{coordinates}

\item{digits}{number of digits to round to}
}
\value{
string with L path component
}
\description{
SVG Path Line
}
\keyword{internal}
