% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{svgMoveTo}
\alias{svgMoveTo}
\title{SVG Path Move}
\usage{
svgMoveTo(x, y, digits = 2)
}
\arguments{
\item{x, y}{coordinates}

\item{digits}{number of digits to round to}
}
\value{
string with M path component
}
\description{
SVG Path Move
}
\examples{
svgMoveTo(20.123, 30.456)
}
