% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{svgQuadraticTo}
\alias{svgQuadraticTo}
\title{SVG Path Quadratic Curve}
\usage{
svgQuadraticTo(x1, y1, x, y, digits = 2)
}
\arguments{
\item{x1, y1}{coordinates of control point}

\item{x, y}{coordinates of main point}

\item{digits}{number of digits to round to}
}
\value{
string with Q path component
}
\description{
SVG Path Quadratic Curve
}
\examples{
svgQuadraticTo(20.123, 30.456, 40.123, 50.456)
}
