% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{tem_def}
\alias{tem_def}
\title{Get template definition}
\usage{
tem_def(tem_id = "frl", path = NULL)
}
\arguments{
\item{tem_id}{numeric ID of template to retrieve from webmorph.org}

\item{path}{path of local tem definition file}
}
\value{
list with template definition
}
\description{
Template definitions are specially formatted files that contain information about templates that are needed to do things like symmetrising and masking images. This function is mostly used internally.
}
\examples{
frl <- tem_def("frl")
frl$points[1:10, ]

fpp106 <- tem_def("fpp106")
fpp106$lines |> str()

fpp83 <- tem_def("fpp83")
fpp83$mask |> str()

}
