% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stim.R
\name{tem_text}
\alias{tem_text}
\title{Make text version of a template}
\usage{
tem_text(stim)
}
\arguments{
\item{stim}{A list of class stim (one item in a stimlist)}
}
\value{
The text for a .tem file
}
\description{
Make text version of a template
}
\examples{
stimuli <- demo_stim()
f_tem_text <- tem_text(stimuli$f_multi)
}
\keyword{internal}
