% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem_to_xml.R
\name{tem_to_xml}
\alias{tem_to_xml}
\title{Templates to XML}
\usage{
tem_to_xml(stimuli, dir = "images", imageset = "Image Set")
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{dir}{path to save the images and XML file}

\item{image_set}{name for the image set (in the XML file)}
}
\value{
the path to the xml file (invisibly)
}
\description{
Make an XML file with the template points for a set of stimuli. For use with \link{facetrain} for training dlib models.
}
\examples{
\dontrun{
# requires python and dlib
stimuli <- demo_stim("composite") |> 
  subset_tem(features("face"))

# create xml and image directory for training
xml <- tem_to_xml(stimuli)

# read the first 10 lines of the xml file
readLines(xml, n = 10)
}
}
