% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tps.R
\name{tems_to_array}
\alias{tems_to_array}
\title{Convert stimuli to array for geomorph}
\usage{
tems_to_array(stimuli)
}
\arguments{
\item{stimuli}{list of class stimlist}
}
\value{
3D array
}
\description{
Convert stimuli to array for geomorph
}
\examples{
data <- demo_stim() \%>\% tems_to_array()
dim(data)

}
