% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_size.R
\name{to_size}
\alias{to_size}
\title{Resize and crop images to a specified size}
\usage{
to_size(
  stimuli,
  width,
  height = NULL,
  fill = wm_opts("fill"),
  patch = FALSE,
  squash = FALSE,
  keep_rels = FALSE
)
}
\arguments{
\item{stimuli}{of class stimlist}

\item{width}{the target width (or a vector of width and height)}

\item{height}{the target height (or null if width is dimensions)}

\item{fill}{background color if cropping goes outside the original image}

\item{patch}{whether to use the patch function to set the background color}

\item{squash}{whether to move template points outside the image boundaries inside the image}

\item{keep_rels}{whether to keep the size relationships between images in the set, or make all the maximum size}
}
\value{
stimlist with cropped tems and/or images
}
\description{
Resize and crop images to a specified size
}
\examples{

stimuli <- c(
  demo_stim("lisa")[[1]],
  demo_stim("composite")[[1]],
  demo_stim("rainbow")[[1]]
)

to_size(stimuli, 300, 400, fill = "red") \%>\% plot()
to_size(stimuli, 300, 400, fill = "red", keep_rels = TRUE) \%>\% plot()


}
