% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimlist.R
\name{validate_stimlist}
\alias{validate_stimlist}
\title{Validate a list of stimuli}
\usage{
validate_stimlist(x, tem = FALSE)
}
\arguments{
\item{x}{The object}

\item{tem}{require templates}
}
\value{
A stimlist
}
\description{
Check if an object is a list of class stimlist. If class stim, wrap in stimlist. If a properly structured list without the right class, add the right class. If the img is not a magick_image or the pointer is dead, reloads from the imgpath.
}
