% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg.R
\name{webmorph_up}
\alias{webmorph_up}
\title{Check if webmorph.org is available}
\usage{
webmorph_up()
}
\description{
Check if webmorph.org is available
}
\examples{
webmorph_up()
}
\seealso{
WebMorph.org functions
\code{\link{avg}()},
\code{\link{continuum}()},
\code{\link{loop}()},
\code{\link{symmetrize}()},
\code{\link{trans}()}
}
\concept{webmorph}
