% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmorphR-package.R
\docType{package}
\name{webmorphR}
\alias{webmorphR}
\alias{webmorphR-package}
\title{webmorphR: Reproducible Face Stimuli}
\description{
Create reproducible image stimuli, specialised for images with webmorph.org templates. See https://debruine.github.io/webmorphR/ for detailed tutorials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://debruine.github.io/webmorphR}
  \item \url{https://github.com/debruine/webmorphR}
  \item Report bugs at \url{https://github.com/debruine/webmorphR/issues}
}

}
\author{
\strong{Maintainer}: Lisa DeBruine \email{debruine@gmail.com} (\href{https://orcid.org/0000-0002-7523-5539}{ORCID})

Other contributors:
\itemize{
  \item Emily Cunningham \email{ekc00004@students.stir.ac.uk} [proofreader, contributor]
}

}
\keyword{internal}
