% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{width}
\alias{width}
\title{Image widths}
\usage{
width(stimuli, type = c("all", "min", "max", "unique"))
}
\arguments{
\item{stimuli}{list of stimuli}

\item{type}{whether to return all widths, min, max, or only unique widths}
}
\value{
vector of widths
}
\description{
Image widths
}
\examples{

demo_stim() |> width()
}
\seealso{
Other info: 
\code{\link{add_info}()},
\code{\link{compare}()},
\code{\link{get_info}()},
\code{\link{get_point}()},
\code{\link{height}()},
\code{\link{metrics}()},
\code{\link{rename_stim}()}
}
\concept{info}
