% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tps.R
\name{write_tps}
\alias{write_tps}
\title{Create a TPS file from a stimlist}
\usage{
write_tps(stimuli, path_to_tps = NULL)
}
\arguments{
\item{stimuli}{list of stimuli}

\item{path_to_tps}{optional filename to save TPS file}
}
\value{
text of tps file
}
\description{
Create a TPS file from a stimlist
}
\examples{
# set path_to_tps to save to a file
tps <- demo_stim() |>
  write_tps()

}
