% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xget}
\alias{xget}
\title{Get named item when unsure of name}
\usage{
xget(x, ..., .default = NULL)
}
\arguments{
\item{x}{named vector or list}

\item{...}{possible names in x in priority order}

\item{.default}{default value if no names found}
}
\value{
list or vector item
}
\description{
If a user gives you a named vector or list and there are a few possibilities for the item you want to get, list the possible names in order, using integers for indices.
}
\examples{
x <- c(h = 100, w = 200)

width <- xget(x, "width", "w", 1)
height <- xget(x, "height", "h", 2)
}
\keyword{internal}
