% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_delin.R
\name{auto_delin}
\alias{auto_delin}
\title{Auto-Delineation}
\usage{
auto_delin(
  stimuli,
  model = c("dlib7", "dlib70", "fpp106", "fpp83"),
  replace = FALSE,
  face = 1,
  dlib_path = NULL
)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{model}{Which shape predictor model to use (dlib7, dlib70, fpp106, fpp83)}

\item{replace}{if FALSE, only gets templates for images with no template}

\item{face}{which face to delineate in each image if there is more than 1 (only for Face++)}

\item{dlib_path}{path to a custom dlib .dat landmark file to use (model is ignored if set)}
}
\value{
stimlist with templates
}
\description{
Automatically delineate faces using dlib in python or Face++ (an external service). Wrapper function for \code{\link[=fpp_auto_delin]{fpp_auto_delin()}} and \code{\link[webmorphR.dlib:dlib_auto_delin]{webmorphR.dlib::dlib_auto_delin()}}.
}
\examples{
\dontrun{
# requires an API key in .Renviron
  auto_fpp106 <- demo_stim() |>
    auto_delin(model = "fpp106", replace = TRUE)
    
  # requires debruine/webmorphR.dlib
  auto_dlib7 <- demo_stim() |>
    auto_delin(replace = TRUE)
  
  auto_dlib70 <- demo_stim() |>
    auto_delin(model = "dlib70", replace = TRUE)
}
}
\seealso{
\link{fpp_auto_delin}, \code{\link[webmorphR.dlib:dlib_auto_delin]{webmorphR.dlib::dlib_auto_delin()}}
}
