% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_tem.R
\name{average_tem}
\alias{average_tem}
\title{Average templates}
\usage{
average_tem(stimuli, name = "average")
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{name}{Name for the average}
}
\value{
list with class stimlist
}
\description{
This function just averages the templates. An average image is returned, but it is just all the images superimposed. To create a template-aware average, see \code{\link{avg}}.
}
\examples{
demo_stim() \%>\% average_tem()

}
