% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg.R
\name{avg}
\alias{avg}
\title{Average Images}
\usage{
avg(
  stimuli,
  texture = TRUE,
  norm = c("none", "twopoint", "rigid"),
  normpoint = 0:1
)
}
\arguments{
\item{stimuli}{images to average (list of class stimlist)}

\item{texture}{logical, textured average}

\item{norm}{how to normalise}

\item{normpoint}{points for twopoint normalisation}
}
\value{
stimlist with average image and template
}
\description{
Average Images
}
\examples{
\dontrun{
  avg <- demo_stim() |> avg()
}
}
