% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{centroid}
\alias{centroid}
\title{Get center coordinates}
\usage{
centroid(stimuli, points = NULL)
}
\arguments{
\item{stimuli}{list of stimuli}

\item{points}{which points to include (0-based); if NULL, all points will be used}
}
\value{
named matrix of centroid x and y coordinates
}
\description{
Get center coordinates
}
\examples{
demo_stim() |> centroid()

# get the centre of the eye points
demo_stim() |> centroid(0:1)
}
\seealso{
Template functions
\code{\link{auto_delin}()},
\code{\link{average_tem}()},
\code{\link{change_lines}()},
\code{\link{delin}()},
\code{\link{draw_tem}()},
\code{\link{features}()},
\code{\link{get_point}()},
\code{\link{remove_tem}()},
\code{\link{require_tems}()},
\code{\link{same_tems}()},
\code{\link{squash_tem}()},
\code{\link{subset_tem}()},
\code{\link{tem_def}()},
\code{\link{viz_tem_def}()}
}
\concept{tem}
