% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2lab.R
\name{col2lab}
\alias{col2lab}
\title{Color to Lab Conversion}
\usage{
col2lab(col)
}
\arguments{
\item{col}{vector of hex or color names}
}
\value{
list of L, a and b values
}
\description{
R color to LAB colourspace conversion.
}
\examples{
col2lab(c("red", "green", "blue"))

}
