% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{color_conv}
\alias{color_conv}
\title{Convert colors}
\usage{
color_conv(
  color,
  alpha = 1,
  from = c("guess", "col", "hex", "hexa", "hex3", "rgb", "rgba"),
  to = c("hexa", "hex", "rgba", "rgb", "hsv")
)
}
\arguments{
\item{color}{A color in one of the input formats (see Details)}

\item{alpha}{Alpha transparency (values <=1 converted to 0-255); ignored if color has alpha already}

\item{from, to}{Input and output color spaces, see \code{Details} below.}
}
\value{
color in \code{to} format
}
\description{
Convert from common color inputs to specified output type, adding alpha transparency.
}
\details{
color: one of the R colours listed in c\code{colors()}, e.g., "red"
hex: hexadecimal string, e.g., "#FF0000"
hexa: hexadecimal string with alpha, e.g., "#FF0000FF"
hex3: abbreviated hexadecimal string, e.g., "#F00"
rgb: vector of red, green and blue values 0-255, e.g., c(255, 0, 0)
rgb: vector of red, green, blue and alpha values 0-255, e.g., c(255, 0, 0, 255)
hsv: vector of hue, saturation and value values (0-1), e.g., c(h=0, s = 1, v = 1)
}
\examples{
color_conv("red")
color_conv("#FF0000")
color_conv("#FF0000FF")
color_conv(c(255,0,0))
color_conv(c(255,0,0,255))

color_conv("dodgerblue", 0.5, to = "rgba")

}
