% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{continuum}
\alias{continuum}
\title{Continuum}
\usage{
continuum(from_img, to_img, from = 0, to = 1, by = 0.1, ...)
}
\arguments{
\item{from_img}{image to start at}

\item{to_img}{image to end at}

\item{from}{starting percentage}

\item{to}{ending percentage}

\item{by}{step size}

\item{...}{arguments to pass to \code{\link{trans}}}
}
\value{
stimlist
}
\description{
Continuum
}
\examples{
\dontrun{
  stimuli <- demo_stim()
  cont <- continuum(stimuli$f_multi, stimuli$m_multi)
  animate(cont, 20, rev = TRUE)
}
}
