% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tem.R
\name{convert_tem}
\alias{convert_tem}
\title{Convert templates}
\usage{
convert_tem(
  stimuli,
  from = c("guess", "frl", "fpp106", "fpp83", "dlib70", "dlib7"),
  to = c("frl", "fpp106", "fpp83", "dlib70", "dlib7")
)
}
\arguments{
\item{stimuli}{list of stimuli}

\item{from}{id of template definition of stimlist images}

\item{to}{id of template definition to convert to}
}
\value{
list of stimuli with converted templates
}
\description{
Convert templates
}
\keyword{internal}
