% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_stim.R
\name{demo_stim}
\alias{demo_stim}
\title{Demo Stimuli}
\usage{
demo_stim(set = c("test", "tem_examples"), pattern = NULL)
}
\arguments{
\item{set}{the images set ("test" or "tem_examples")}

\item{pattern}{defaults to all files}
}
\value{
stimlist
}
\description{
A convenience function to get demo stimuli
}
\examples{
demo_stim() |> plot()

}
