% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{get_point}
\alias{get_point}
\title{Get Point Coordinates}
\usage{
get_point(stimuli, pt = 0)
}
\arguments{
\item{stimuli}{list of class stimlist with templates}

\item{pt}{point(s) to return}
}
\value{
data frame of x and y coordinates of the specified point(s) for each stimulus
}
\description{
Get a data frame of the x and y coordinates of a template point
}
\examples{
demo_stim() \%>\% get_point(0:1)
}
