% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_func.R
\name{greyscale}
\alias{greyscale}
\alias{grayscale}
\title{Make images greyscale}
\usage{
greyscale(stimuli)

grayscale(stimuli)
}
\arguments{
\item{stimuli}{list of class stimuli}
}
\value{
stimlist with new images
}
\description{
Make images greyscale
}
\examples{
demo_stim() |> greyscale() |> plot()
demo_stim() |> grayscale()
}
