% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{height}
\alias{height}
\title{Image heights}
\usage{
height(stimuli, type = c("all", "min", "max", "unique"))
}
\arguments{
\item{stimuli}{list of stimuli}

\item{type}{whether to return all heights, min, max, or only unique heights}
}
\value{
vector of heights
}
\description{
Image heights
}
\examples{

demo_stim() |> height()
}
\seealso{
Other info: 
\code{\link{add_info}()},
\code{\link{compare}()},
\code{\link{get_info}()},
\code{\link{get_point}()},
\code{\link{metrics}()},
\code{\link{rename_stim}()},
\code{\link{width}()}
}
\concept{info}
