% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Label images}
\usage{
label(stimuli, ...)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{...}{arguments to pass on to \code{mlabel} or \code{gglabel}}
}
\value{
stimlist with labelled images
}
\description{
Defaults to \link{mlabel} unless you use arguments specific to \link{gglabel}.
}
\examples{
stimuli <- demo_stim("test")
m_stimuli <- stimuli |>
  label(text = c("CHINWE", "GEORGE"), 
        gravity = c("north", "south"),
        color = "red")
        
gg_stimuli <- stimuli |>
  label(label = c("CHINWE", "GEORGE"), 
        x = 0.5, 
        y = c(1, 0.02),
        vjust = c(1, 0), 
        size = 10,
        color = "red")
        
# c(stimuli, gg_stimuli, m_stimuli) |> plot(nrow = 3)
}
