% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Label images}
\usage{
label(stimuli, ...)
}
\arguments{
\item{stimuli}{list of stimuli}

\item{...}{arguments to pass on to \code{mlabel} or \code{gglabel}}
}
\value{
stimlist with labelled images
}
\description{
Defaults to \link{mlabel} unless you use arguments specific to \link{gglabel}. All arguments are vectorised over the stimuli and values are recycled or truncated if there are fewer or more than stimuli.
}
\examples{
stimuli <- demo_stim()

# label with magick::image_annotate
m_stimuli <- label(stimuli, 
                   text = c("CHINWE", "GEORGE"), 
                   gravity = c("north", "south"),
                   color = "red")

# label with ggplot2::annotate
gg_stimuli <- label(stimuli,
                    label = c("CHINWE", "GEORGE"), 
                    x = 0.5, 
                    y = c(1, 0.02),
                    vjust = c(1, 0), 
                    size = 10,
                    color = "red")

}
\seealso{
\code{\link[=mlabel]{mlabel()}}, \code{\link[=gglabel]{gglabel()}}
}
