% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_tbl.R
\name{list_to_tbl}
\alias{list_to_tbl}
\title{List format to table format}
\usage{
list_to_tbl(list, rownames = NULL)
}
\arguments{
\item{list}{a list of lists}

\item{rownames}{whether to return a table with no rownames (NULL), rownames from the list item names (NA), or as a new column (the column name as a string)}
}
\value{
a data table
}
\description{
List format to table format
}
\examples{

list_format <- list(
  item1 = list(x = 1, y = "A"),
  item2 = list(x = 2),
  item3 = list(y = "C", z = TRUE),
  item4 = list(x = NULL, y = NULL, z = NULL),
  item5 = list(),
  item6 = NULL
)

list_to_tbl(list_format, rownames = "item")

}
