% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimlist.R
\name{new_stim}
\alias{new_stim}
\title{Make a new stim}
\usage{
new_stim(img, path = "", ...)
}
\arguments{
\item{img}{Image made by magick}

\item{path}{Path to image file (or name)}

\item{...}{Additional items for stim}
}
\value{
list with class stim
}
\description{
Make a new stim
}
\seealso{
Other stim: 
\code{\link{as_stimlist}()},
\code{\link{new_stimlist}()},
\code{\link{read_stim}()},
\code{\link{write_stim}()}
}
\concept{stim}
