% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimlist.R
\name{new_stimlist}
\alias{new_stimlist}
\title{Make a new stimlist}
\usage{
new_stimlist(..., .names = NULL)
}
\arguments{
\item{...}{Lists with class "stim"}

\item{.names}{Names for each stimulus}
}
\value{
A stimlist
}
\description{
Make a new stimlist
}
