% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{pad}
\alias{pad}
\title{Pad images}
\usage{
pad(stimuli, top = 10, right = top, bottom = top, left = right, ...)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{top}{number of pixels to pad the top}

\item{right}{number of pixels to pad the right side}

\item{bottom}{number of pixels to pad the bottom}

\item{left}{number of pixels to pad the left side}

\item{...}{additional arguments to pass to \code{\link{crop}}}
}
\value{
list of class stimlist
}
\description{
Convenience function to calcuate offsets for crop
}
\examples{
padded <- demo_stim() |> pad(10, fill = "red")
plot(padded)
}
