% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch}
\alias{patch}
\title{Patch colour}
\usage{
patch(
  img,
  x1 = 1,
  x2 = 10,
  y1 = 1,
  y2 = 10,
  color = c("hex", "rgb"),
  func = stats::median
)
}
\arguments{
\item{img}{The image}

\item{x1}{starting x pixel of patch}

\item{x2}{ending x pixel of patch}

\item{y1}{starting y pixel of patch}

\item{y2}{ending y pixel of patch}

\item{color}{The type of color to return (hex, rgb)}

\item{func}{The function to apply to an array of rgba values to determine the central colour (defaults to median)}
}
\value{
hex or rgba color value
}
\description{
Get the median (or mean or user-defined function) colour value of a specified patch of pixels on an image. This is useful for matching background colours.
}
\examples{
patch(demo_stim()[[1]]$img)

}
