% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch}
\alias{patch}
\title{Patch colour}
\usage{
patch(
  stimuli,
  x1 = 0,
  x2 = 10,
  y1 = 0,
  y2 = 10,
  color = c("hex", "rgb"),
  func = stats::median
)
}
\arguments{
\item{stimuli}{list of stimuli}

\item{x1, x2, y1, y2}{start and end pixels of the patch, if <=1, interpreted as proportions of width or height}

\item{color}{The type of color to return (hex, rgb)}

\item{func}{The function to apply to an array of rgba values to determine the central colour (defaults to median, but mean, min, or max are also useful)}
}
\value{
a vector of hex or rgba color values
}
\description{
Get the median (or mean or user-defined function) colour value of a specified patch of pixels on an image. This is useful for matching background colours.
}
\details{
The colour values of each pixel in the patch are converted to CIE-Lab values before using the func to calculate the central tendency of the L (lightness), a (red-green axis) and b (blue-yellow axis); see \code{\link[=col2lab]{col2lab()}} and \code{\link[=lab2rgb]{lab2rgb()}} for more details.

This excludes transparent pixels, and returns "transparent" if all pixels in the patch are transparent.
}
\examples{
# get colour from the upper left corder
demo_stim() |> patch()

# get median colour from centre .1 width pixels
demo_stim() |> patch(x1 = .45, x2 = .55, y1 = .45, y2 = .55)

# get mean rgb colour from a 10-pixel strip across the top of the image
demo_stim() |> patch(0, 1, 0, 10, color = "rgb", func = mean)

}
