% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.stim}
\alias{plot.stim}
\title{Plot stim}
\usage{
\method{plot}{stim}(x, y, ...)
}
\arguments{
\item{x}{stim}

\item{y}{omitted}

\item{...}{Arguments to be passed to \code{\link[=plot_stim]{plot_stim()}}}
}
\value{
stimlist
}
\description{
Plot stim
}
\seealso{
Visualisation functions
\code{\link{as_ggplot}()},
\code{\link{draw_tem}()},
\code{\link{gglabel}()},
\code{\link{label}()},
\code{\link{mlabel}()},
\code{\link{plot.stimlist}()},
\code{\link{plot_rows}()},
\code{\link{plot_stim}()}
}
\concept{viz}
\keyword{internal}
