% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{procrustes_align}
\alias{procrustes_align}
\title{Procrustes align templates}
\usage{
procrustes_align(data, rotate = 0)
}
\arguments{
\item{data}{Template points}

\item{rotate}{Whether to "guess" the rotation, or rotate 0, 90, 180, or 270 degrees}
}
\value{
coordinates
}
\description{
Procrustes align templates
}
\keyword{internal}
