% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stim.R
\name{read_stim}
\alias{read_stim}
\title{Read stimuli}
\usage{
read_stim(path, pattern = NULL, ...)
}
\arguments{
\item{path}{Path to directory containing image and/or template files (or a single file path)}

\item{pattern}{Vector of patterns to use to search for files, or a vector of image indices (e.g., 1:4 selects the first 4 images and their templates if they exist)}

\item{...}{further arguments to pass on to other functions, e.g., breaks argument for unique_names()}
}
\value{
a list with class stimlist
}
\description{
Read stimuli
}
\examples{
path <- system.file("extdata/composite", package = "webmorphR")
stimuli <- read_stim(path, "multi")

}
