% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{remove_tem}
\alias{remove_tem}
\title{Remove templates}
\usage{
remove_tem(stimuli)
}
\arguments{
\item{stimuli}{list of class stimlist}
}
\value{
list of class stimlist
}
\description{
Remove templates
}
\examples{
demo_stim() \%>\% remove_tem()
}
