% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate templates and images}
\usage{
rotate(
  stimuli,
  degrees = 0,
  fill = wm_opts("fill"),
  patch = FALSE,
  keep_size = TRUE
)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{degrees}{degrees to rotate}

\item{fill}{background color}

\item{patch}{whether to use the patch function to set the background color}

\item{keep_size}{whether to keep the original size or expand images to the new rotated size}
}
\value{
stimlist with rotated tems and/or images
}
\description{
Rotate templates and images
}
\examples{
rotated <- demo_stim() |>
  rotate(45, fill = "dodgerblue") |>
  draw_tem()

rotate_patch <- demo_stim() |>
  rotate(45, patch = TRUE, keep_size = FALSE)

}
