% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{same_tems}
\alias{same_tems}
\title{Check All Templates are the Same}
\usage{
same_tems(stimuli)
}
\arguments{
\item{stimuli}{list of stimuli}
}
\value{
logical
}
\description{
Check All Templates are the Same
}
\examples{
stim <- demo_stim()
stim2 <- subset_tem(stim, features("gmm"))

same_tems(stim)

c(stim, stim2) |> same_tems()
}
\seealso{
Template functions
\code{\link{auto_delin}()},
\code{\link{average_tem}()},
\code{\link{centroid}()},
\code{\link{change_lines}()},
\code{\link{delin}()},
\code{\link{draw_tem}()},
\code{\link{features}()},
\code{\link{get_point}()},
\code{\link{remove_tem}()},
\code{\link{squash_tem}()},
\code{\link{subset_tem}()},
\code{\link{tem_def}()},
\code{\link{viz_tem_def}()}
}
\concept{tem}
