% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{same_tems}
\alias{same_tems}
\title{Check All Templates are the Same}
\usage{
same_tems(stimuli)
}
\arguments{
\item{stimuli}{list of class stimlist}
}
\value{
logical
}
\description{
Check All Templates are the Same
}
\examples{
demo_stim() \%>\% same_tems()
c(demo_stim("test"), demo_stim("lisa")) \%>\% same_tems()
}
