% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{svgBezier}
\alias{svgBezier}
\title{Construct Bezier Curves from Template Points}
\usage{
svgBezier(v, idx = 1)
}
\arguments{
\item{v}{matrix of template points with x and y as rows and point as columns}

\item{idx}{index of line segment within the path}
}
\value{
string with path component
}
\description{
Construct Bezier Curves from Template Points
}
\examples{
stimuli <- demo_stim()
# get upper and lower eye points
l_upper <- stimuli$f_multi$points[, 18:22]
l_lower <- stimuli$f_multi$points[, c(22, 30, 29, 28, 18)]
svgBezier(l_upper, 1)
svgBezier(l_lower, 2)
}
