% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{symmetrize}
\alias{symmetrize}
\title{Symmetrise Images}
\usage{
symmetrize(stimuli, shape = 1, color = 1, tem_id = "FRL", ...)
}
\arguments{
\item{stimuli}{images to symmatrise (list of class stimlist)}

\item{shape, color}{amount of symmetry (0 for none, 1.0 for perfect)}

\item{tem_id}{template ID to be passed to \code{\link{tem_def}} (usually "frl" or "fpp106")}

\item{...}{Additional arguments to pass to \code{\link{trans}}}
}
\value{
stimlist with symmatrised images and templates
}
\description{
Symmetrise Images
}
\examples{
\dontrun{
  stimuli <- demo_stim("london", 1) \%>\% resize(0.5)

  sym_both <- symmetrize(stimuli)
  sym_shape <- symmetrize(stimuli, color = 0)
  sym_color <- symmetrize(stimuli, shape = 0)
  sym_anti <- symmetrize(stimuli, shape = -1.0, color = 0)
}
}
