% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{tem_to_xml}
\alias{tem_to_xml}
\title{Templates to XML}
\usage{
tem_to_xml(stimuli, dir = "images")
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{dir}{path to save XML file and images}
}
\description{
Make an XML file with the template points for a set of stimuli. For use with training dlib.
}
