% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_size.R
\name{to_size}
\alias{to_size}
\title{Resize and crop/pad images to a specified size}
\usage{
to_size(
  stimuli,
  width,
  height = NULL,
  fill = wm_opts("fill"),
  patch = FALSE,
  crop = FALSE,
  keep_rels = FALSE
)
}
\arguments{
\item{stimuli}{of class stimlist}

\item{width}{the target width (or a vector of width and height)}

\item{height}{the target height (or null if width is dimensions)}

\item{fill}{background color if cropping goes outside the original image}

\item{patch}{whether to use the patch function to set the background color}

\item{crop}{whether to crop or pad images to make them the specified size}

\item{keep_rels}{whether to keep the size relationships between images in the set, or make all the maximum size}
}
\value{
stimlist with cropped tems and/or images
}
\description{
Resize and crop/pad images to a specified size
}
\examples{

# make images with different aspect ratios and sizes
stimuli <- demo_stim() \%>\% crop(c(0.8, 1.0)) \%>\% resize(c(1.0, 0.5))

to_size(stimuli, 300, 400, fill = "dodgerblue") \%>\% plot()
to_size(stimuli, 300, 400, fill = "dodgerblue", keep_rels = TRUE) \%>\% plot()

}
