% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\title{Transform Images}
\usage{
trans(
  trans_img = NULL,
  from_img = NULL,
  to_img = NULL,
  shape = 0,
  color = 0,
  texture = 0,
  outname = NULL,
  norm = c("none", "twopoint", "rigid"),
  normpoint = 0:1,
  sample_contours = TRUE,
  warp = c("multiscale", "linear", "multiscalerb")
)
}
\arguments{
\item{trans_img}{list of stimuli to transform}

\item{from_img}{negative transform dimension endpoint (0\% image)}

\item{to_img}{positive transform dimension endpoint (100\% image)}

\item{shape, color, texture}{amount to change along the vector defined by from_img and to_img (can range from -3 to +3)}

\item{outname}{name to save each image as}

\item{norm}{how to normalise the images}

\item{normpoint}{points for twopoint normalisation}

\item{sample_contours}{whether to sample contours or just points}

\item{warp}{warp type}
}
\value{
list of stimuli with transformed images and templates
}
\description{
Transform Images
}
\examples{
\dontrun{
  stimuli <- demo_stim()
  sexdim <- trans(stimuli, stimuli$f_multi, stimuli$m_multi,
                  shape = c(fem = -0.5, masc = 0.5))
  sexdim |> draw_tem() |> label() |> plot()
}

}
\seealso{
WebMorph.org functions
\code{\link{avg}()},
\code{\link{continuum}()},
\code{\link{loop}()},
\code{\link{symmetrize}()}
}
\concept{webmorph}
