% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{viz_tem_def}
\alias{viz_tem_def}
\title{Visualise a template definition}
\usage{
viz_tem_def(tem_def, ...)
}
\arguments{
\item{tem_def}{the template definition; usually from \code{\link[=tem_def]{tem_def()}}}

\item{...}{further arguments to pass to \code{\link[=draw_tem]{draw_tem()}}; pt.size and line.size often need to be adjusted}
}
\value{
a stimlist with a blank image and the template drawn on it
}
\description{
Visualise a template definition
}
\examples{
dlib70 <- tem_def("dlib70")
viz_tem_def(dlib70, pt.size = 5, line.size = 3)
}
\seealso{
Template functions
\code{\link{auto_delin}()},
\code{\link{average_tem}()},
\code{\link{centroid}()},
\code{\link{change_lines}()},
\code{\link{delin}()},
\code{\link{draw_tem}()},
\code{\link{features}()},
\code{\link{get_point}()},
\code{\link{remove_tem}()},
\code{\link{require_tems}()},
\code{\link{same_tems}()},
\code{\link{squash_tem}()},
\code{\link{subset_tem}()},
\code{\link{tem_def}()}
}
\concept{tem}
