% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmorphR-package.R
\docType{package}
\name{webmorphR}
\alias{webmorphR}
\title{webmorphR: Reproducible Face Stimuli}
\description{
Create reproducible image stimuli, specialised for images with webmorph.org templates. See https://debruine.github.io/webmorphR/ for detailed tutorials.
}
\keyword{package}
