% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webmorphR-package.R
\docType{package}
\name{webmorphR}
\alias{webmorphR}
\title{webmorphR: Reproducible Face Stimuli}
\description{
Create reproducible face stimuli using images with webmorph.org templates. \url{https://facelab.github.io/webmorphR/}
}
\keyword{package}
