% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{width}
\alias{width}
\title{Image widths}
\usage{
width(stimuli, type = c("all", "min", "max", "unique"))
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{type}{whether to return all widths, min, max, or only unique widths}
}
\value{
vector of widths
}
\description{
Image widths
}
\examples{

demo_stim() |> width()
}
