% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stim.R
\name{write_stim}
\alias{write_stim}
\title{Write tems and images to files}
\usage{
write_stim(stimuli, dir = ".", format = "png", ...)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{dir}{Directory to save to}

\item{format}{output format such as "png", "jpeg", "gif"}

\item{...}{other arguments to pass to magick::image_write}
}
\value{
list of saved paths
}
\description{
Write tems and images to files
}
\examples{
\dontrun{
  demo_stim() \%>\% write_stim("test_faces", "jpg")
}
}
