% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stim.R
\name{write_stim}
\alias{write_stim}
\title{Write tems and images to files}
\usage{
write_stim(
  stimuli,
  dir = ".",
  names = NULL,
  format = "png",
  ...,
  ask = interactive(),
  overwrite = !ask
)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{dir}{Directory to save to}

\item{names}{A vector of stimulus names or NULL to use names fro the stimuli list}

\item{format}{output format such as "png", "jpeg", "gif"}

\item{...}{other arguments to pass to magick::image_write}

\item{ask}{ask if you want to overwrite existing files}

\item{overwrite}{whether to overwrite existing files}
}
\value{
list of saved paths
}
\description{
Write tems and images to files
}
\examples{
\dontrun{
  demo_stim() \%>\% write_stim("test_faces", format = "jpg")
}
}
